+++ projects/ids-enterprise-ng/src/lib/splitter/soho-splitter.component.ts
@@ -80,6 +80,15 @@ export class SohoSplitterComponent implements AfterViewInit, OnDestroy {
     }
   }
 
+  /** Ability to stop dragging at a max left or right size. **/
+  @Input() set maxWidth(maxWidth: SohoSplitterOptionsMaxWidth) {
+    this.options.maxWidth = maxWidth;
+    if (this.splitter) {
+      this.splitter.settings.maxWidth = maxWidth;
+      this.splitter.updated();
+    }
+  }
+
   // Determines if the split percentage should be saved to local storage
   @Input() set save(save: boolean) {
     this.options.save = save;